adding a serious amount of new examples, seriously, then I fixed the test, then I fixed Laura's updates, let's see what happens
